+++ visit_manager/postgres_utils/models/misc.py
@@ -9,4 +9,10 @@ class VisitStatus(str, enum.Enum):
     cancelled = "cancelled"
 
 
-from visit_manager.postgres_utils.models import PaymentStatus
+class PaymentStatus(str, enum.Enum):
+    __tablename__ = "payment_status"
+    processing = "processing"
+    cancelled = "cancelled"
+    error = "error"
+    success = "success"
+    refunded = "refunded"
